requirements.txt
attrs==21.2.0
boto3==1.17.98
botocore==1.20.98
certifi==2021.5.30
chardet==4.0.0
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
iniconfig==1.1.1
instabot==0.117.0
jmespath==0.10.0
mock==4.0.3
moviepy==1.0.3
numpy==1.21.0
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
proglog==0.1.9
py==1.10.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.13.3
s3transfer==0.4.2
schedule==1.1.0
six==1.16.0
toml==0.10.2
tqdm==4.61.1
urllib3==1.26.5
